+++ src/components/SearchBox/SearchBox.tsx
@@ -1,30 +1,26 @@
-import React, { useRef, useState } from 'react';
-import styled from 'styled-components';
-import throttle from 'lodash/throttle';
-import SearchIcon from '@material-ui/icons/Search';
-import Paper from '@material-ui/core/Paper';
-import IconButton from '@material-ui/core/IconButton';
-import Router from 'next/router';
-import match from 'autosuggest-highlight/match';
+import React, { useRef, useState } from "react";
+import styled from "styled-components";
+import throttle from "lodash/throttle";
+import SearchIcon from "@material-ui/icons/Search";
+import Paper from "@material-ui/core/Paper";
+import IconButton from "@material-ui/core/IconButton";
+import Router from "next/router";
+import match from "autosuggest-highlight/match";
 
-import { fetchJson } from '../../services/fetch';
-import { useMapStateContext } from '../utils/MapStateContext';
-import { useFeatureContext } from '../utils/FeatureContext';
-import { AutocompleteInput } from './AutocompleteInput';
-import { t } from '../../services/intl';
-import { ClosePanelButton } from '../utils/ClosePanelButton';
-import { isDesktop, useMobileMode } from '../helpers';
-import { getGlobalMap } from '../../services/mapStorage';
-import { Feature, LineString, Point } from '../../services/types';
-import { getPoiClass } from '../../services/getPoiClass';
-import { getCenter } from '../../services/getCenter';
-import { OsmApiId } from '../../services/helpers';
-import { presets } from '../../services/tagging/data';
+import { fetchJson } from "../../services/fetch";
+import { useMapStateContext } from "../utils/MapStateContext";
+import { useFeatureContext } from "../utils/FeatureContext";
+import { AutocompleteInput } from "./AutocompleteInput";
+import { t } from "../../services/intl";
+import { ClosePanelButton } from "../utils/ClosePanelButton";
+import { isDesktop, useMobileMode } from "../helpers";
+import { presets } from "../../services/tagging/data";
 import {
   fetchSchemaTranslations,
   getPresetTermsTranslation,
-  getPresetTranslation,
-} from '../../services/tagging/translations';
+  getPresetTranslation
+} from "../../services/tagging/translations";
+import { performOverpassSearch } from "../../services/overpassSearch";
 
 const TopPanel = styled.div`
   position: absolute;
@@ -65,104 +61,6 @@ const getApiUrl = (inputValue, view) => {
   const lvl = Math.max(0, Math.min(16, Math.round(zoom)));
   const q = encodeURIComponent(inputValue);
   return `https://photon.komoot.io/api/?q=${q}&lon=${lon}&lat=${lat}&zoom=${lvl}`;
-};
-
-const overpassQuery = (bbox) => `[out:json][timeout:25];
-(
-  node["sport"="climbing"](${bbox});
-  way["sport"="climbing"](${bbox});
-  relation["sport"="climbing"](${bbox});
-  node["climbing"="route_bottom"](${bbox});
-  way["climbing"="route_bottom"](${bbox});
-  relation["climbing"="route_bottom"](${bbox});
-);
-out body;
->;
-out skel qt;`;
-
-const getOverpassUrl = ([a, b, c, d]) =>
-  `https://overpass-api.de/api/interpreter?data=${encodeURIComponent(
-    overpassQuery([d, a, b, c]),
-  )}`;
-
-const notNull = (x) => x != null;
-
-const getGeometry = {
-  node: ({ lat, lon }): Point => ({ type: 'Point', coordinates: [lon, lat] }),
-  way: (foo): LineString => {
-    const { geometry } = foo;
-    return {
-      type: 'LineString',
-      coordinates: geometry?.filter(notNull)?.map(({ lat, lon }) => [lon, lat]),
-    };
-  },
-  relation: ({ members }): LineString => ({
-    type: 'LineString',
-    coordinates: members[0]?.geometry
-      ?.filter(notNull)
-      ?.map(({ lat, lon }) => [lon, lat]),
-  }),
-};
-
-export const overpassAroundToSkeletons = (response: any): Feature[] =>
-  response.elements.map((element) => {
-    const { type, id, tags = {} } = element;
-    const geometry = getGeometry[type]?.(element);
-    return {
-      type: 'Feature',
-      osmMeta: { type, id },
-      tags,
-      properties: { ...getPoiClass(tags), tags },
-      geometry,
-      center: getCenter(geometry) ?? undefined,
-    };
-  });
-
-// maybe take inspiration from https://github.com/tyrasd/osmtogeojson/blob/gh-pages/index.js
-const osmJsonToSkeletons = (response: any): Feature[] => {
-  const nodesById = response.elements
-    .filter((element) => element.type === 'node')
-    .reduce((acc, node) => {
-      acc[node.id] = node;
-      return acc;
-    }, {});
-
-  const getGeometry2 = {
-    node: ({ lat, lon }): Point => ({ type: 'Point', coordinates: [lon, lat] }),
-    way: (way): LineString => {
-      const { nodes } = way;
-      return {
-        type: 'LineString',
-        coordinates: nodes
-          ?.map((nodeId) => nodesById[nodeId])
-          .map(({ lat, lon }) => [lon, lat]),
-      };
-    },
-    relation: ({ members }): LineString => ({
-      type: 'LineString',
-      coordinates: members[0]?.geometry
-        ?.filter(notNull)
-        ?.map(({ lat, lon }) => [lon, lat]),
-    }),
-  };
-
-  return response.elements.map((element) => {
-    const { type, id, tags = {} } = element;
-    const geometry = getGeometry2[type]?.(element);
-    return {
-      type: 'Feature',
-      osmMeta: { type, id },
-      tags,
-      properties: { ...getPoiClass(tags), ...tags },
-      geometry,
-      center: getCenter(geometry) ?? undefined,
-    };
-  });
-};
-
-export const convertOsmIdToMapId = (apiId: OsmApiId) => {
-  const osmToMapType = { node: 0, way: 1, relation: 4 };
-  return parseInt(`${apiId.id}${osmToMapType[apiId.type]}`, 10);
 };
 
 // https://docs.mapbox.com/help/troubleshooting/working-with-large-geojson-data/
@@ -221,21 +119,7 @@ const findInPresets = (inputValue) => {
 
 const fetchOptions = throttle(async (inputValue, view, setOptions, bbox) => {
   if (inputValue === 'climbing') {
-    const overpass = await fetchJson(getOverpassUrl(bbox));
-    console.log(overpass);
-    const map = getGlobalMap();
-    // put overpass features on mapbox gl map
-    const features = osmJsonToSkeletons(overpass)
-      .filter(
-        (feature) => feature.center && Object.keys(feature.tags).length > 0,
-      )
-      .map((feature) => ({
-        ...feature,
-        id: convertOsmIdToMapId(feature.osmMeta),
-      }));
-    console.log('overpass geojson', features);
-    map.getSource('overpass')?.setData({ type: 'FeatureCollection', features });
-
+    await performOverpassSearch(bbox);
     setOptions([]);
     return;
   }
